Rollback of warnings fix which breaks RPC
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@26407 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/tapi/tapi_cfg.c
@@ -1304,7 +1304,7 @@ tapi_cfg_neigh_op(enum tapi_cfg_oper op, const char *ta,
 int
 tapi_cfg_get_hwaddr(const char *ta,
                     const char *ifname,
-                    void *hwaddr, size_t *hwaddr_len)
+                    void *hwaddr, unsigned int *hwaddr_len)
 {
     char     buf[1024];
     int      rc;
